Coerce frame position to integral values

+++ app/src/main/java/com/gaurav/avnc/ui/vnc/FrameState.kt
@@ -288,8 +288,8 @@ class FrameState(
         if (safeArea.isEmpty || !safeArea.intersect(0f, 0f, vpWidth, vpHeight))
             safeArea.set(0f, 0f, vpWidth, vpHeight)
 
-        frameX = coercePosition(frameX, safeArea.left, safeArea.right, fbWidth)
-        frameY = coercePosition(frameY, safeArea.top, safeArea.bottom, fbHeight)
+        frameX = coercePosition(frameX, safeArea.left, safeArea.right, fbWidth).toInt().toFloat()
+        frameY = coercePosition(frameY, safeArea.top, safeArea.bottom, fbHeight).toInt().toFloat()
     }
 
     /**
